Remove use of deprecated imp module (#37)
`imp` has been deprecated since python 3.4. The documentation[1]
suggests to use the `ModuleType` constructor instead, which is what I'm
doing here.

The use of the deprecated `imp` module caused a deperecation warning,
which then caused a doctest failure in sage on nixos. Not sure why it
apparently wasn't an issue with the python3 swich on
sage-the-distribution.

[1] https://docs.python.org/3.5/library/imp.html#imp.new_module